Move common tests from `AppGeneratorTest` and `PluginGeneratorTest` to `SharedGeneratorTests`
