+++ railties/lib/rails/generators/rails/plugin/templates/test/test_helper.rb
@@ -17,6 +17,7 @@ Minitest.backtrace_filter = Minitest::BacktraceFilter.new
 Rails::TestUnitReporter.executable = 'bin/test'
 <% end -%>
 
+<% unless options[:skip_active_record] -%>
 # Load fixtures from the engine
 if ActiveSupport::TestCase.respond_to?(:fixture_path=)
   ActiveSupport::TestCase.fixture_path = File.expand_path("fixtures", __dir__)
@@ -24,3 +25,4 @@ if ActiveSupport::TestCase.respond_to?(:fixture_path=)
   ActiveSupport::TestCase.file_fixture_path = ActiveSupport::TestCase.fixture_path + "/files"
   ActiveSupport::TestCase.fixtures :all
 end
+<% end -%>
